+++ src/i18/messages.ts
@@ -24,7 +24,7 @@ export interface Message {
 	SCOPE_TYPE_CUSTOM: string;
 	SCOPE_TAG_VALUE_INPUT_PLACEHOLDER: string;
 	SCOPE_EXTENSION_VALUE_INPUT_PLACEHOLDER: string;
-	SCOPE_SPCIFIC_FOLDER_INPUT_PLACEHOLDER: string;
+	SCOPE_SPECIFIC_FOLDER_INPUT_PLACEHOLDER: string;
 
 	// file name format
 	FILE_NAME_FORMAT_LABEL: string;
