+++ src/manager/constants.ts
@@ -1,3 +1,14 @@
 export const DEFAULT_DATETIME_FORMAT = "yyyyMMddHHmmss";
 
 export const SAMPLE_DATE = new Date('2024-12-31 23:59:59')
+
+/** 视为图片的扩展名（linkGenerator 与附件库共用） */
+export const IMAGE_EXTENSIONS: string[] = [
+	"png",
+	"jpg",
+	"jpeg",
+	"gif",
+	"svg",
+	"webp",
+	"bmp",
+];
